# fix: mark Hash arg as noescape