--- a/utils.go
+++ b/utils.go
@@ -102,3 +102,15 @@ func mulFold64(x, y u64) u64 {
 	hi, lo := bits.Mul64(x, y)
 	return hi ^ lo
 }
+
+// noescape hides a pointer from escape analysis.  noescape is
+// the identity function but escape analysis doesn't think the
+// output depends on the input. noescape is inlined and currently
+// compiles down to zero instructions.
+// USE CAREFULLY!
+// This was copied from the runtime; see issues 23382 and 7921.
+//go:nosplit
+func noescape(p unsafe.Pointer) unsafe.Pointer {
+	x := uintptr(p)
+	return unsafe.Pointer(x ^ 0)
+}