Go code that uses `unsafe.Pointer` requires this hack or `[]byte` arg always escapes to heap. [cespare/xxhash](https://github.com/cespare/xxhash/blob/v2.1.2/xxhash_amd64.go#L10) uses asm with `go:noescape` instead. 

See https://github.com/golang/go/issues/23382 for an example from stdlib

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkHashEscape-12     27.1          4.61          -83.03%

benchmark                  old allocs     new allocs     delta
BenchmarkHashEscape-12     1              0              -100.00%

benchmark                  old bytes     new bytes     delta
BenchmarkHashEscape-12     8             0             -100.00%
```