@zeebo 2022-02-04
Can you demonstrate with a test case or benchmark how this leaks the argument? Note, the Hash and HashString methods are carefully designed to be inlined by the compiler.