@vmihailenco 2022-02-11
I've added a benchmark. PTAL