@vmihailenco 2022-02-16
>How about adding it to HashSeed, Hash128 and Hash128Seed?

@zeebo what do you think?