@vcabbage 2022-03-04
To my understanding, this change appears actually unsafe. It's intentionally breaking the [`unsafe.Pointer`](https://pkg.go.dev/unsafe#Pointer) rules, specifically rule 3, to hide a pointer. In https://go.dev/cl/86976/ the resulting pointer is only used for the copy check and is never dereferenced.

I'll note that the escape seems to be due to the function indirection and can be avoided by restructuring as below. Though it's unclear to me whether the indirection is itself an optimization.

```go
func Hash(b []byte) uint64 {
	if len(b) <= 16 {
		return hashSmall(*(*str)(ptr(&b)))
	}
	return hashMed(*(*str)(ptr(&b)))
}
```