@zeebo 2022-03-04
I'm playing around with removing the call through the function pointer as well. One thing I'm going to experiment with is in go1.18, switches are potentially compiled differently as jump tables sometimes. Maybe some of the losses can be recouped there.