@vmihailenco 2022-03-05
>It's intentionally breaking the [unsafe.Pointer](https://pkg.go.dev/unsafe#Pointer) rules

My understanding is that this change can't interfere with unsafe rules, because it does not change the generated code in this package. It only tweaks escape analysis on the calling side of `Hash` func to not allocate the passed arg on the heap. If the arg is not used/referenced by xxh3 after `Hash` function exits, this change should be safe.