@vcabbage 2022-03-05
I'm not an authority on this, so take my thoughts with the appropriate grain of salt 🙂

> My understanding is that this change can't interfere with unsafe rules, because it does not change the generated code in this package.

The `noescape` function doesn't generate any code directly, but changing escape analysis can change the generated code. Heap allocating or not is part of the code.

> It only hacks escape analysis on the calling side of Hash func to not allocate the passed arg on the heap. If the arg is not used/referenced by xxh3 after Hash function exits, this change should be safe.

If there were no more references to `b` in the calling code, could it end up being collected while the hash function is executing? I don't know with any certainty, but with the unsafe rules broken it wouldn't surprise me.

---

Apparently the usage in `strings.Builder` has [caused an issue](https://github.com/golang/go/issues/47276#issuecomment-882598415). I imagine the experienced folks that used it in `strings.Builder` weren't expecting that, highlighting that the `USE CAREFULLY!` comment shouldn't be overlooked.

At the very least, if it's possible to avoid the allocation without using such a tricky function that seems preferable.