@zeebo 2022-03-05
So I experimented quite a bit with shuffling the functions around to avoid the call through the pointer, and here's a table summarizing the results:

<summary>
<details>

size           | 64      | 64 seed | 128     | 128 seed
----           | --      | ------- | ---     | --------
0              | -6.79%  | ~       | -7.37%  | ~
1              | +16.13% | +6.59%  | +6.16%  | +0.13%
2              | +16.51% | +6.26%  | +6.02%  | +5.19%
3              | +20.22% | +7.20%  | +8.84%  | +2.49%
4              | +8.36%  | +5.62%  | +8.23%  | +2.47%
8              | +8.63%  | +5.36%  | +8.18%  | +2.50%
9              | +7.67%  | +3.54%  | +5.13%  | +6.04%
16             | +7.70%  | +4.29%  | +5.07%  | +5.99%
17             | -14.78% | -5.22%  | -2.67%  | -2.50%
32             | -14.78% | -5.12%  | -4.10%  | -3.07%
33             | -7.71%  | -0.28%  | -3.61%  | -5.13%
64             | -7.66%  | ~       | -3.64%  | -4.94%
65             | -1.84%  | +1.93%  | -3.36%  | -3.98%
96             | -2.06%  | +2.09%  | -4.21%  | -4.14%
97             | -1.11%  | +1.99%  | -3.66%  | -7.57%
128            | -1.39%  | +1.98%  | -5.44%  | -5.47%
129            | -7.96%  | +2.09%  | -6.50%  | -7.87%
240            | -7.21%  | -1.12%  | -5.80%  | -6.59%
241            | -10.54% | -10.31% | -11.72% | -11.94%
241-AVX2       | -5.55%  | -9.49%  | -5.37%  | -11.21%
241-SSE2       | -5.63%  | -9.81%  | -4.93%  | -9.89%
512            | -14.08% | -11.17% | -12.21% | -10.57%
512-AVX2       | -4.99%  | -9.42%  | -5.16%  | -8.47%
512-SSE2       | -4.51%  | -8.58%  | -3.14%  | -6.90%
1024           | -13.21% | -10.98% | -13.44% | -11.05%
1024-AVX2      | -3.41%  | -8.51%  | -2.62%  | -7.92%
1024-SSE2      | -3.06%  | -8.32%  | -2.60%  | -5.57%
8192           | -9.90%  | -10.41% | -12.90% | -11.01%
8192-AVX2      | -1.96%  | -2.78%  | -0.67%  | -1.68%
8192-SSE2      | -0.48%  | -1.92%  | -0.93%  | -1.86%
102400         | -5.89%  | -6.81%  | -10.05% | -6.98%
102400-AVX2    | -1.74%  | -1.95%  | -1.59%  | -0.87%
102400-SSE2    | -3.77%  | -0.62%  | +1.17%  | -0.59%
1024000        | -5.62%  | -6.46%  | -8.96%  | -5.89%
1024000-AVX2   | -1.41%  | -1.55%  | +0.96%  | +1.23%
1024000-SSE2   | -3.07%  | -1.22%  | +3.10%  | +2.82%
10240000       | -4.24%  | -6.19%  | -9.74%  | -7.33%
10240000-AVX2  | -1.86%  | +6.80%  | -5.89%  | ~
10240000-SSE2  | ~       | ~       | -3.93%  | -4.88%
102400000      | -4.29%  | -5.04%  | -10.29% | -6.98%
102400000-AVX2 | -8.43%  | -11.67% | ~       | -5.40%
102400000-SSE2 | -3.56%  | -2.71%  | -5.17%  | -5.29%

</details>
</summary>

There's a lot to like about that: sizes above 16 are pretty much entirely wins, sometimes on the order of 15%. On the other hand, sizes smaller than 16 are sometimes worse. But while the percents are large, in absolute terms, they account for less than 1ns. Like, the size 3 that gained 20% went from 2.8ns/op to 3.35ns/op. 

So I think this is acceptable, and I'm gonna go with it rather than the "noescape" trick. As noted above, the output code being the same doesn't imply the runtime metadata the garbage collector uses is, and the fact that it can be attributed to causing some invalid GC state makes me even more wary.

Thanks for noticing this and providing the test case to demonstrate it.