@vmihailenco 2022-03-06
Thanks for fixing this.

Hard to say if this would help with performance, but I wonder why do you convert everything to `str` and not use `[]byte` as is? That would remove unsafe entirely for the common case of hashing bytes.

For strings, you could use unsafe to cast `string` to `[]byte`.