@zeebo 2022-03-06
The code used to pass two parameters: the data pointer and length. It was changed to pass the `str` struct because that made the inliner happy. Now that the body of the functions are simpler again, it could go back to doing more "normal" unsafe stuff, but there is no difference.

It would be incorrect to use unsafe to interpret a `string` as a `[]byte`. A `[]byte` is larger than a `string`: it has a data pointer, a length and a capacity, whereas the `string` only has the data pointer and length. 

`str` has the same memory layout as a `string` and the same memory layout as the first two fields of a `[]byte`, so unsafe casting both to the same thing allows sharing the same code to process both.