@vmihailenco 2022-03-06
>Now that the body of the functions are simpler again

I meant to use `[]byte` and restore optimization for small sizes. Use `[]byte`, no unsafe, and no indirection.