@vmihailenco 2022-03-06
>It would be incorrect to use unsafe to interpret a string as a []byte

Every second library has a function to cast `string` to `[]byte` using unsafe. Perhaps you mean `slower` instead of `incorrect`.

```go
// Bytes converts string to byte slice.
func Bytes(s string) []byte {
	return *(*[]byte)(unsafe.Pointer(
		&struct {
			string
			Cap int
		}{s, len(s)},
	))
}
```